Half-Done destructring with routes
